+++ config/initializers/schked.rb
@@ -1,3 +1,9 @@
 # frozen_string_literal: true
 
-Schked.config.paths << Sbmt::Outbox::Engine.root.join("config", "schedule.rb")
+begin
+  require "schked"
+
+  Schked.config.paths << Sbmt::Outbox::Engine.root.join("config", "schedule.rb")
+rescue LoadError
+  # optional dependency
+end